remove compiler warnings from descriptors
Remove the following gcc warnings:
-Wunused-variable
-Wconversion
-Wsign-conversion

Reviewed-on: http://git.am.freescale.net:8181/10639
Tested-by: Cristian Stoica <cristian.stoica@freescale.com>
(cherry picked from commit 90a501b0650f630e0d4796f140e4f48871c85593)
Reviewed-on: http://git.am.freescale.net:8181/10688

+++ sec/rta/include/flib/jobdesc.h
@@ -57,12 +57,12 @@ static inline void cnstr_jobdesc_mdsplitkey(uint32_t *descbuf,
 		64,	/* SHA384 */
 		64	/* SHA512 */
 	};
-	uint8_t split_key_len, idx;
+	uint32_t split_key_len, idx;
 	struct program prg;
 	struct program *program = &prg;
 
 	idx = (cipher & OP_ALG_ALGSEL_SUBMASK) >> OP_ALG_ALGSEL_SHIFT;
-	split_key_len = mdpadlen[idx] * 2;
+	split_key_len = (uint32_t)(mdpadlen[idx] * 2);
 
 	PROGRAM_CNTXT_INIT(descbuf, 0);
 	if (ps)
